requirements.txt
pip~=21.3.1
wheel~=0.37.1
setuptools~=60.2.0
pyglet~=2.0.10
darkdetect~=0.8.0
customtkinter~=5.2.1
pillow~=10.2.0
pygame~=2.5.2